Create search page and move login to index
