--- src/pages/login.vue
@@ -1,16 +1,0 @@
-<script setup lang="ts">
-import { useAuthStore } from '@/stores/auth'
-
-const authStore = useAuthStore()
-
-async function AuthorizeUser() {
-  const authUrl = await authStore.generateUserAuthorizationUrl()
-  window.location.href = authUrl.toString()
-}
-</script>
-
-<template>
-  <main>
-    <v-btn color="primary" @click="AuthorizeUser" prepend-icon="mdi-spotify" size="large">Login with Spotify</v-btn>
-  </main>
-</template>
